adding the auth middleware
