+++ app/helpers/token.js
@@ -1,5 +1,5 @@
 const moment = require('moment');
-const jwt = require('jwt-simple');
+const jwt = require('jsonwebtoken');
 const { secret } = require('../../config').common.session;
 
 exports.createToken = any => {
@@ -10,5 +10,10 @@ exports.createToken = any => {
       .add(14, 'days')
       .unix()
   };
-  return jwt.encode(payload, secret);
+  return jwt.sign(payload, secret);
+};
+
+exports.decodeToken = token => {
+  const decoded = jwt.verify(token, secret);
+  return decoded;
 };
