+++ app/mappers/users.js
@@ -1,7 +1,7 @@
 const encrypt = require('../services/encrypt');
 
 exports.registerBodyMapper = body => {
-  const hash = encrypt(body.password);
+  const hash = encrypt.encrypt(body.password);
   return {
     ...body,
     lastName: body.last_name,
