+++ app/routes.js
@@ -3,12 +3,13 @@ const { getAlbums, getAlbumById } = require('./controllers/albums');
 const { registerUser, singIn, getUsers } = require('./controllers/users');
 const { registerValidator, sessionValidator } = require('../app/schemas/users');
 const { checkError } = require('./middlewares/validatorSchema');
+const { auth } = require('./middlewares/auth');
 
 exports.init = app => {
   app.get('/health', healthCheck);
   app.get('/albums', getAlbums);
   app.get('/albums/:id/photos', getAlbumById);
-  app.get('/users', getUsers);
+  app.get('/users', auth, getUsers);
   app.post('/users', [registerValidator, checkError], registerUser);
   app.post('/users/sessions', [sessionValidator, checkError], singIn);
 };
