+++ app/services/users.js
@@ -15,17 +15,3 @@ exports.validateUser = user => {
     return Promise.reject(errors.databaseError(error.message));
   });
 };
-
-exports.getUsers = () =>
-  User.findAll({ raw: true })
-    .then(users =>
-      users.map(user => {
-        const us = {
-          name: user.name,
-          last_name: user.last_name,
-          email: user.email
-        };
-        return us;
-      })
-    )
-    .catch(() => Promise.reject(errors.databaseError('cant get users')));
