+++ package.json
@@ -66,9 +66,12 @@
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.2",
     "express": "^4.16.2",
+    "express-paginate": "^1.0.0",
+    "express-pagination-middleware": "^1.0.2",
     "express-validator": "^6.2.0",
     "express-wolox-logger": "1.1.0",
     "factory-girl": "^5.0.4",
+    "jsonwebtoken": "^8.5.1",
     "jwt-simple": "^0.5.1",
     "pg": "^7.4.1",
     "request-promise": "^4.2.4",
